Moved TestRunner to common package, fixed pom.xml issue, reformat code
