+++ pom.xml
@@ -32,12 +32,6 @@
             <artifactId>selenium-chrome-driver</artifactId>
             <version>${chrome.driver.version}</version>
         </dependency>
-        <dependency>
-            <groupId>org.seleniumhq.selenium</groupId>
-            <artifactId>selenium-chrome-driver</artifactId>
-            <version>4.0.0-beta-2</version>
-            <scope>compile</scope>
-        </dependency>
     </dependencies>
 
 </project>
