+++ src/main/java/com/softserveinc/ita/common/TestRunner.java
@@ -1,0 +1,33 @@
+package com.softserveinc.ita.common;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.support.ui.WebDriverWait;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeSuite;
+
+public class TestRunner {
+    private static WebDriver driver;
+    private static WebDriverWait wait;
+
+    public static WebDriver getDriver() {
+        return driver;
+    }
+
+    public static WebDriverWait getWait() {
+        return wait;
+    }
+
+    @BeforeSuite
+    public static void setupWebDriverTypeAndProperties() {
+        System.setProperty("webdriver.chrome.driver", "src/main/resources/drivers/chromedriver.exe");
+        driver = new ChromeDriver();
+        driver.manage().window().maximize();
+        wait = new WebDriverWait(driver, 10);
+    }
+
+    @AfterMethod
+    public void close() {
+        driver.close();
+    }
+}
