+++ src/main/java/com/softserveinc/ita/nromanchuk/GoogleSearchResultPage.java
@@ -1,5 +1,6 @@
 package com.softserveinc.ita.nromanchuk;
 
+import com.softserveinc.ita.common.TestRunner;
 import org.openqa.selenium.By;
 
 import java.util.List;
@@ -8,7 +9,7 @@ import java.util.stream.Collectors;
 public class GoogleSearchResultPage {
 
     public String getLinkText(int index) {
-        return TestRunner.driver.findElements(By.xpath("//h3[@class='LC20lb DKV0Md']"))
+        return TestRunner.getDriver().findElements(By.xpath("//h3[@class='LC20lb DKV0Md']"))
                 .stream()
                 .map(element -> element.getText().toLowerCase())
                 .collect(Collectors.toList()).get(index);
@@ -16,7 +17,7 @@ public class GoogleSearchResultPage {
 
     public List<String> getResultsLinks() {
         return TestRunner
-                .driver
+                .getDriver()
                 .findElements(By.xpath("//div[@class='yuRUbf']/a"))
                 .stream()
                 .map(webElement -> webElement.getAttribute("href"))
