--- src/main/java/com/softserveinc/ita/nromanchuk/TestRunner.java
@@ -1,24 +1,0 @@
-package com.softserveinc.ita.nromanchuk;
-
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeSuite;
-
-public class TestRunner {
-    public static WebDriver driver;
-
-    @BeforeSuite
-    public void setup() {
-        System.setProperty("webdriver.chrome.driver", "drivers//chromedriver.exe");
-        driver = new ChromeDriver();
-        driver.manage().window().maximize();
-    }
-
-    @AfterMethod
-    public void close() {
-        if (TestRunner.driver != null) {
-            TestRunner.driver.quit();
-        }
-    }
-}
