+++ src/main/java/com/softserveinc/ita/sbendus/pageobject/GoogleSearchPage.java
@@ -1,5 +1,6 @@
 package com.softserveinc.ita.sbendus.pageobject;
 
+import com.softserveinc.ita.common.TestRunner;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
