+++ src/main/java/com/softserveinc/ita/shladkyi/GoogleHomePage.java
@@ -1,6 +1,7 @@
 package com.softserveinc.ita.shladkyi;
 
 
+import com.softserveinc.ita.common.TestRunner;
 import org.openqa.selenium.By;
 import org.openqa.selenium.Keys;
 
@@ -8,12 +9,12 @@ import org.openqa.selenium.Keys;
 public class GoogleHomePage {
 
     public GoogleHomePage open() {
-        TestRunner.driver.get("https://www.google.com/");
+        TestRunner.getDriver().get("https://www.google.com/");
         return this;
     }
 
     public GoogleSearchResultPage searchFor(String searchTerm) {
-        TestRunner.driver
+        TestRunner.getDriver()
                 .findElement(By.cssSelector("[class='gLFyf gsfi']"))
                 .sendKeys(searchTerm, Keys.ENTER);
         return new GoogleSearchResultPage();
