+++ src/main/java/com/softserveinc/ita/shladkyi/GoogleSearchResultPage.java
@@ -1,5 +1,6 @@
 package com.softserveinc.ita.shladkyi;
 
+import com.softserveinc.ita.common.TestRunner;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 
@@ -8,12 +9,12 @@ import java.util.stream.Collectors;
 
 public class GoogleSearchResultPage {
     public String getSearchResultLinkText(int indexOfLink) {
-        List<WebElement> allSearchResultLinks = TestRunner.driver.findElements(By.xpath("//h3[@class='LC20lb DKV0Md']"));
+        List<WebElement> allSearchResultLinks = TestRunner.getDriver().findElements(By.xpath("//h3[@class='LC20lb DKV0Md']"));
         return allSearchResultLinks.get(indexOfLink).getText();
     }
 
     public List<String> getAllSearchResultLinks() {
-        List<WebElement> allSearchResultLinks = TestRunner.driver.findElements(By.xpath("//div[@class='TbwUpd NJjxre']"));
+        List<WebElement> allSearchResultLinks = TestRunner.getDriver().findElements(By.xpath("//div[@class='TbwUpd NJjxre']"));
         return allSearchResultLinks.stream().map(WebElement::getText).collect(Collectors.toList());
     }
 }
