--- src/main/java/com/softserveinc/ita/shladkyi/TestRunner.java
@@ -1,16 +1,0 @@
-package com.softserveinc.ita.shladkyi;
-
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.testng.annotations.BeforeSuite;
-
-public class TestRunner {
-    public static WebDriver driver;
-
-    @BeforeSuite
-    public void setUp() {
-        System.setProperty("webdriver.chrome.driver", "C:/Users/Inspiron/Downloads/chromedriver_win32/chromedriver.exe");
-        driver = new ChromeDriver();
-        driver.manage().window().maximize();
-    }
-}
